Update e3e64bb1bea2ab58575b849dfef6fd51.txt
Using $ tr -s ' ' < outputfile.txt to squeeze the double whitespaces out which came about by replacing newlines with whitespace. The use of double whitespace shall be reserved for format preservation such as paragraph and perhaps triple for pagealthough digitally remastered do not need pages as that would only arise for printing and would be handled by the printing software.